app/protobufs/api_gateway/requirements.txt
grpc-interceptor >= 0.11.0
grpcio-tools >= 1.30
pymongo >= 3.11.2
dnspython >= 2.1.0
connexion[swagger-ui]
connexion >= 2.0.0
flask >= 0.0.0
prometheus_client

pyopenssl >= 0.0.0
flask_httpauth >= 0.0.0
flask_cors >= 0.0.0
itsdangerous >= 0.0.0